pyproject.toml
[tool.poetry]
name = "nlp-summarization"
packages = [{include = "src"}]

[project]
name = "nlp-summarization"
version = "0.1.0"
description = "Dialogue Summarization with KoBART/KoT5 Ensemble"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.7.1",
    "transformers>=4.54.0",
    "pytorch-lightning>=2.5.2",
    "accelerate>=0.26.0",
    "wandb>=0.21.0",
    "rouge>=1.0.1",
    "pandas>=2.3.1",
    "tqdm>=4.67.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "openai>=1.97.1",
    "scikit-learn>=1.3.0",
    "seaborn>=0.13.0",
    "matplotlib>=3.8.0",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
